refactor(home, styles): lit migration
