+++ src/storybook/pages/home/home.stories.tsx
@@ -13,8 +13,12 @@ import {
   wrapperStyle,
 } from './home.common';
 import './home.scss';
-import type { Meta, StoryObj, ArgTypes, Args } from '@storybook/html';
+import type { Meta, StoryObj, ArgTypes, Args } from '@storybook/web-components';
 import type { InputType } from '@storybook/types';
+
+import './home.common';
+import '../../../components/sbb-alert';
+import '../../../components/sbb-alert-group';
 
 /* ************************************************* */
 /* Storybook controls                                */
