+++ src/storybook/styles/layout/layout.stories.tsx
@@ -2,7 +2,8 @@
 import { Fragment, h, JSX } from 'jsx-dom';
 import readme from './readme.md?raw';
 import './layout.scss';
-import type { Meta, StoryObj } from '@storybook/html';
+import type { Meta, StoryObj } from '@storybook/web-components';
+import '../../../components/sbb-link';
 
 const PageSpacingTemplate = (): JSX.Element => (
   <section class="sbb-page-spacing visualized-page-spacing">
